Move tests for TransferTokensActionToParticleGroupsMapper to separate common test file, preparing for shared logic with coming BurnTokensActionToParticleGroupsMapper tests
